backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
groq>=0.4.0
deepgram-sdk==3.5.0
httpx[http2]>=0.25.2
websockets>=13.0
twilio==8.10.0
python-multipart==0.0.9
aiofiles==23.2.1
fpdf
requests>=2.31.0
